fix: fixed bug in verification file

+++ test/yosupo/shortest_path.test.cpp
@@ -4,7 +4,7 @@
 int main(){
   int n,m,s,t;
   cin>>n>>m>>s>>t;
-  Graph<int>g(n);
+  Graph<long long>g(n);
   g.read(m,0,true,true);
   auto d=dijkstra_path(g,s,t);
   if(d.first==-1){
